requirements.txt
# api
fastapi==0.115.8
uvicorn==0.34.0
spacy==3.8.3
transformers==4.48.3
torch==2.6.0
PyYAML==6.0.2
pydantic==2.10.6
nltk==3.9.1
scikit-learn==1.6.1
python-multipart

# evaluation 
llama-index
openai
httpx
python-dotenv